+++ src/components/review/ReviewTable.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import useCheckMobileScreen from '../../../../services/checkMobileScreen'
+import useCheckMobileScreen from '../../services/checkMobileScreen'
 import ReviewDesktopTable from './ReviewDesktopTable'
 import ReviewMobileTable from './ReviewMobileTable'
