+++ src/containers/Gurb/components/TextRecomendation.jsx
@@ -2,7 +2,7 @@ import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 
 import { textHeader2, textBody1 } from '../gurbTheme'
-import RequiredTitle from './InputTitle'
+import RequiredTitle from '../../../components/InputTitle'
 
 const TextRecomendation = ({ title, text, required = false }) => {
   return (
