+++ src/containers/Gurb/gurbTheme.js
@@ -129,15 +129,6 @@ export const containerSpacing = {
   borderRadius: '8px'
 }
 
-export const iconRequirements = {
-  borderRadius: '8px',
-  border: '1px solid #D9D9D9',
-  color: '#1E1E1E',
-  fontSize: '3rem',
-  padding: '8px',
-  backgroundColor: '#fff'
-}
-
 export const iconOffRequirements = {
   borderRadius: '8px',
   border: '1px solid #D9D9D9',
