+++ src/containers/Gurb/pages/Contract/HolderPersonalData.jsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next'
 
-import InputField from '../../components/InputField'
+import InputField from '../../../../components/InputField'
 
 import Grid from '@mui/material/Grid'
