+++ src/themes/commonStyles.js
@@ -1,0 +1,9 @@
+
+export const iconRequirements = {
+  borderRadius: '8px',
+  border: '1px solid #D9D9D9',
+  color: '#1E1E1E',
+  fontSize: '3rem',
+  padding: '8px',
+  backgroundColor: '#fff'
+}
